+++ clean_arch_linter/lib/src/models/clean_arch_linter_config.dart
@@ -1,14 +1,23 @@
 import 'package:analysis_server_plugin_core/analysis_server_plugin_core.dart';
-import 'package:clean_arch_linter/src/models/ddr_config.dart';
+import 'package:clean_arch_linter/src/models/cross_layer_import_config.dart';
+import 'package:clean_arch_linter/src/models/dart_sdk_import_config.dart';
+import 'package:clean_arch_linter/src/models/domain_config.dart';
+import 'package:clean_arch_linter/src/models/third_party_import_config.dart';
 
 class CleanArchLinterConfig extends ContextConfig {
-  final DependencyDirectionRuleConfig ddrConfig;
+  final DomainConfig domainConfig;
+  final DartSDKImportConfig dartSDKConfig;
+  final CrossLayerImportConfig crossLayerConfig;
+  final ThirdPartyImportConfig thirdPartyConfig;
 
   CleanArchLinterConfig({
     required super.packageInfo,
     required super.logConfig,
     super.scanConfig,
-    this.ddrConfig = const DependencyDirectionRuleConfig(),
+    this.domainConfig = const DomainConfig(),
+    this.dartSDKConfig = const DartSDKImportConfig(),
+    this.crossLayerConfig = const CrossLayerImportConfig(),
+    this.thirdPartyConfig = const ThirdPartyImportConfig(),
   });
 
   @override
@@ -16,6 +25,9 @@ class CleanArchLinterConfig extends ContextConfig {
     'packageInfo': packageInfo.toMap(),
     'logConfig': logConfig.toMap(),
     'scanConfig': scanConfig.toMap(),
-    'ddrConfig': ddrConfig.toMap(),
+    'domainConfig': domainConfig.toMap(),
+    'dartSDKConfig': dartSDKConfig.toMap(),
+    'crossLayerConfig': crossLayerConfig.toMap(),
+    'thirdPartyConfig': thirdPartyConfig.toMap(),
   };
 }