+++ clean_arch_linter/lib/src/models/cross_layer_import_config.dart
@@ -1,0 +1,9 @@
+class CrossLayerImportConfig {
+  final List<String> excludedProjectPaths;
+
+  const CrossLayerImportConfig({this.excludedProjectPaths = const []});
+
+  Map<String, dynamic> toMap() => {
+    'excludedProjectPaths': excludedProjectPaths,
+  };
+}