+++ clean_arch_linter/lib/src/models/dart_sdk_import_config.dart
@@ -1,0 +1,9 @@
+class DartSDKImportConfig {
+  final List<String> excludedDartPackages;
+
+  const DartSDKImportConfig({this.excludedDartPackages = const []});
+
+  Map<String, dynamic> toMap() => {
+    'excludedDartPackages': excludedDartPackages,
+  };
+}