+++ clean_arch_linter/lib/src/models/third_party_import_config.dart
@@ -1,0 +1,9 @@
+class ThirdPartyImportConfig {
+  final List<String> excludedLibraryPackages;
+
+  const ThirdPartyImportConfig({this.excludedLibraryPackages = const []});
+
+  Map<String, dynamic> toMap() => {
+    'excludedLibraryPackages': excludedLibraryPackages,
+  };
+}